+++ src/main/java/pe/edu/upc/Karwas/model/repository/RoleRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.Karwas.model.entity.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 	List<Role> findByNameRole( String nameRole ) throws Exception;
 }
